CU-8692ng0gu_Create-Jenkins-pipeline-for-ORB_SLAM-build-test_Guy-Rodoy (#3)

+++ android_build/build_android.sh
@@ -11,6 +11,7 @@ ANDROID_PLATFORM="30"
 EIGEN_VERSION="3.4.0"
 MAKE="make -j4"
 SKIP_DEPS=0
+COPY_SLAM_ARTIFACTS_TO_APP="true"
 
 # Enable errexit option
 set -e -o pipefail
@@ -27,7 +28,7 @@ trap 'handle_error' ERR
 # help menu
 function script_help {
     echo "Usage: ./android_build/build_android.sh [parameters]"
-    echo -e "  --app_path\t\t\tAbsolute path to DemoApp root"
+    echo -e "  --app_path\t\t\t<optional>Absolute path to DemoApp root. If not mentioned, slam artifacts wont copied to DemoApp"
     echo -e "  --ndk_path\t\t<optional> Absolute path to NDK (default is /home/${USER}/Android/Sdk/ndk/25.1.8937393)"
     echo -e "  --abi\t\t<optional> android abi to build for (arm64-v8a, armeabi-v7a, x86, x86_64). default is arm64-v8a"
     echo -e "  --platform\t\t<optional> android api level to build for (default is 30)"
@@ -62,9 +63,7 @@ done
 
 # Check if required parameters are provided
 if [ -z "$APP_PATH" ]; then
-    echo -e "${RED}Error: The --app_path parameter is required.${DEFAULT_COLOR}"
-    script_help
-    exit 0
+    COPY_SLAM_ARTIFACTS_TO_APP="false"
 fi
 
 case "$ANDROID_ABI" in
@@ -149,47 +148,47 @@ cd build
 $MAKE
 cd $SLAM_ROOT
 
-echo "Copying Shared Objects"
-JNI_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/jniLibs/$ANDROID_ABI
-mkdir -p $JNI_PATH
-cp -r -f -v -a Thirdparty/g2o/lib/libg2o.so $JNI_PATH
-cp -r -f -v -a Thirdparty/openssl/libcrypto.so $JNI_PATH
-cp -r -f -v -a Thirdparty/openssl/libcrypto.so.1.1 $JNI_PATH
-cp -r -f -v -a lib/* $JNI_PATH
+if [ $COPY_SLAM_ARTIFACTS_TO_APP == "true" ]; then
+    echo "Copying Shared Objects"
+    JNI_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/jniLibs/$ANDROID_ABI
+    mkdir -p $JNI_PATH
+    cp -r -f -v -a Thirdparty/g2o/lib/libg2o.so $JNI_PATH
+    cp -r -f -v -a Thirdparty/openssl/libcrypto.so $JNI_PATH
+    cp -r -f -v -a Thirdparty/openssl/libcrypto.so.1.1 $JNI_PATH
+    cp -r -f -v -a lib/* $JNI_PATH
+    echo "Copying Headers"
+    APP_HEADERS_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/jni
+    SLAM_HEADERS_PATH=$APP_HEADERS_PATH/SLAM
+    mkdir -p $SLAM_HEADERS_PATH
+    cp -r -f -v include/* $SLAM_HEADERS_PATH
+    DBOW2_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/DBoW2
+    mkdir -p $DBOW2_HEADERS_PATH
+    cp -r -f -v Thirdparty/DBoW2/DBoW2 Thirdparty/DBoW2/DUtils  $DBOW2_HEADERS_PATH
+    G2O_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/g2o
+    mkdir -p $G2O_HEADERS_PATH
+    cp -r -f -v Thirdparty/g2o/g2o Thirdparty/g2o/config.h $G2O_HEADERS_PATH
+    SOPHUS_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/Sophus
+    mkdir -p $SOPHUS_HEADERS_PATH
+    cp -r -f -v Thirdparty/Sophus/sophus $SOPHUS_HEADERS_PATH
+    EIGEN_HEADERS_PATH=$APP_HEADERS_PATH
+    mkdir -p $EIGEN_HEADERS_PATH
+    cp -r -f -v Thirdparty/eigen-$EIGEN_VERSION/Eigen $EIGEN_HEADERS_PATH
+    BOOST_HEADERS_PATH=$APP_HEADERS_PATH
+    mkdir -p $BOOST_HEADERS_PATH
+    cp -r -f -v Thirdparty/Boost-for-Android/boost_1_82_0/boost $BOOST_HEADERS_PATH
 
-echo "Copying Headers"
-APP_HEADERS_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/jni
-SLAM_HEADERS_PATH=$APP_HEADERS_PATH/SLAM
-mkdir -p $SLAM_HEADERS_PATH
-cp -r -f -v include/* $SLAM_HEADERS_PATH
-DBOW2_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/DBoW2
-mkdir -p $DBOW2_HEADERS_PATH
-cp -r -f -v Thirdparty/DBoW2/DBoW2 Thirdparty/DBoW2/DUtils  $DBOW2_HEADERS_PATH
-G2O_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/g2o
-mkdir -p $G2O_HEADERS_PATH
-cp -r -f -v Thirdparty/g2o/g2o Thirdparty/g2o/config.h $G2O_HEADERS_PATH
-SOPHUS_HEADERS_PATH=$SLAM_HEADERS_PATH/Thirdparty/Sophus
-mkdir -p $SOPHUS_HEADERS_PATH
-cp -r -f -v Thirdparty/Sophus/sophus $SOPHUS_HEADERS_PATH
-EIGEN_HEADERS_PATH=$APP_HEADERS_PATH
-mkdir -p $EIGEN_HEADERS_PATH
-cp -r -f -v Thirdparty/eigen-$EIGEN_VERSION/Eigen $EIGEN_HEADERS_PATH
-BOOST_HEADERS_PATH=$APP_HEADERS_PATH
-mkdir -p $BOOST_HEADERS_PATH
-cp -r -f -v Thirdparty/Boost-for-Android/boost_1_82_0/boost $BOOST_HEADERS_PATH
+    echo "Extracting and copying ORB_SLAM3 assets"
+    APP_ASSETS_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/assets
+    tar -xzf "Vocabulary/ORBvoc.txt.tar.gz" --directory $SLAM_ROOT/android_build/assets/ORB3_SLAM || { echo "Error: Failed to extract ORBvoc.txt"; }
+    cp -r -f -v android_build/assets/ORB3_SLAM $APP_ASSETS_PATH
+    rm -f $SLAM_ROOT/android_build/assets/ORB3_SLAM/ORBvoc.txt
 
-echo "Extracting and copying ORB_SLAM3 assets"
-APP_ASSETS_PATH=$APP_PATH/ServiceApp/com.LibbaInc.ltd/app/src/main/assets
-tar -xzf "Vocabulary/ORBvoc.txt.tar.gz" --directory $SLAM_ROOT/android_build/assets/ORB3_SLAM || { echo "Error: Failed to extract ORBvoc.txt"; }
-cp -r -f -v android_build/assets/ORB3_SLAM $APP_ASSETS_PATH
-rm -f $SLAM_ROOT/android_build/assets/ORB3_SLAM/ORBvoc.txt
-
-echo "Copying OpenCV"
-OLD_VER_STR="VERSION_1_6"
-NEW_VER_STR="VERSION_1_8"
-OPENCV_GRADLE_PATH="${SLAM_ROOT}/Thirdparty/OpenCV-android-sdk/sdk/build.gradle"
-# Search and replace in file:
-sed -i "s/${OLD_VER_STR}/${NEW_VER_STR}/g" "$OPENCV_GRADLE_PATH"
-cp -r -f -v Thirdparty/OpenCV-android-sdk $APP_PATH/ServiceApp
-
+    echo "Copying OpenCV"
+    OLD_VER_STR="VERSION_1_6"
+    NEW_VER_STR="VERSION_1_8"
+    OPENCV_GRADLE_PATH="${SLAM_ROOT}/Thirdparty/OpenCV-android-sdk/sdk/build.gradle"
+    # Search and replace in file:
+    sed -i "s/${OLD_VER_STR}/${NEW_VER_STR}/g" "$OPENCV_GRADLE_PATH"
+    cp -r -f -v Thirdparty/OpenCV-android-sdk $APP_PATH/ServiceApp
+fi
 echo -e "${GREEN}SUCCESS${DEFAULT_COLOR}"
